@@ -30,24 +30,17 @@
   "publishConfig": {
     "access": "public"
   },
-  "peerDependencies": {
-    "@angular/core": "^5.2.9",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "rxjs": "^5.5.8"
-  },
+  "peerDependencies": {},
   "devDependencies": {
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
+    "@angular-devkit/schematics": "^0.6.8",
+    "@angular/common": "6.0.4",
+    "@angular/compiler": "6.0.4",
+    "@angular/compiler-cli": "6.0.4",
+    "@angular/core": "6.0.4",
+    "@angular/forms": "6.0.4",
+    "@angular/http": "6.0.4",
+    "@angular/platform-browser": "6.0.4",
+    "@angular/platform-browser-dynamic": "6.0.4",
     "@types/bootstrap": "^3.3.38",
     "@types/ckeditor": "0.0.45",
     "@types/jasmine": "~2.5.53",
@@ -94,10 +87,10 @@
     "rollup-plugin-node-resolve": "^3.0.3",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "rollup-plugin-uglify": "^3.0.0",
-    "rxjs": "5.5.5",
+    "rxjs": "6.2.0",
     "semantic-release": "^12.4.1",
     "systemjs": "0.19.39",
-    "typescript": "~2.4.2",
-    "zone.js": "0.8.18"
+    "typescript": "~2.7.2",
+    "zone.js": "0.8.26"
   }
 }
